pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xmcp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastmcp>=2.5.0,<3.0",
  "httpx",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "pytest-httpx",
  "ruff",
]
grok = [
  "xai-sdk",
  "xdk",
]

[tool.setuptools]
py-modules = ["server"]
packages = ["auth", "xmcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
